changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11784 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Data Processing/Histogram Objects/HistogramMultiPlot/ORMultiPlot.h
@@ -56,8 +56,8 @@
 
 #pragma mark ¥¥¥Data Source Methods
 - (NSMutableArray*) rois:(int)index;
-- (unsigned)  count;
-- (unsigned)  cachedCount;
+- (NSUInteger)  count;
+- (NSUInteger)  cachedCount;
 - (id)   name;
 @end
